+++ pydantic/__init__.py
@@ -4,7 +4,7 @@ from .env_settings import BaseSettings
 from .error_wrappers import ValidationError
 from .errors import *
 from .fields import Required
-from .main import BaseConfig, BaseModel, create_model, validate_model
+from .main import BaseConfig, BaseModel, Extra, create_model, validate_model
 from .parse import Protocol
 from .types import *
 from .version import VERSION