+++ tests/test_main.py
@@ -3,7 +3,7 @@ from typing import Any, ClassVar, List
 
 import pytest
 
-from pydantic import BaseModel, NoneBytes, NoneStr, Required, ValidationError, constr
+from pydantic import BaseModel, Extra, NoneBytes, NoneStr, Required, ValidationError, constr
 
 
 def test_success():
@@ -135,30 +135,6 @@ def test_recursion_fails():
         RecursiveModel(grape=1, banana=123)
 
 
-class PreventExtraModel(BaseModel):
-    foo = 'whatever'
-
-    class Config:
-        ignore_extra = False
-
-
-def test_prevent_extra_success():
-    m = PreventExtraModel()
-    assert m.foo == 'whatever'
-
-    m = PreventExtraModel(foo=1)
-    assert m.foo == '1'
-
-
-def test_prevent_extra_fails():
-    with pytest.raises(ValidationError) as exc_info:
-        PreventExtraModel(foo='ok', bar='wrong', spam='xx')
-    assert exc_info.value.errors() == [
-        {'loc': ('bar',), 'msg': 'extra fields not permitted', 'type': 'value_error.extra'},
-        {'loc': ('spam',), 'msg': 'extra fields not permitted', 'type': 'value_error.extra'},
-    ]
-
-
 def test_not_required():
     class Model(BaseModel):
         a: float = None
@@ -184,9 +160,77 @@ def test_allow_extra():
         a: float = ...
 
         class Config:
-            allow_extra = True
+            extra = Extra.allow
 
     assert Model(a='10.2', b=12).dict() == {'a': 10.2, 'b': 12}
+
+
+def test_forbidden_extra_success():
+    class ForbiddenExtra(BaseModel):
+        foo = 'whatever'
+
+        class Config:
+            extra = Extra.forbid
+
+    m = ForbiddenExtra()
+    assert m.foo == 'whatever'
+
+    m = ForbiddenExtra(foo=1)
+    assert m.foo == '1'
+
+
+def test_forbidden_extra_fails():
+    class ForbiddenExtra(BaseModel):
+        foo = 'whatever'
+
+        class Config:
+            extra = Extra.forbid
+
+    with pytest.raises(ValidationError) as exc_info:
+        ForbiddenExtra(foo='ok', bar='wrong', spam='xx')
+    assert exc_info.value.errors() == [
+        {'loc': ('bar',), 'msg': 'extra fields not permitted', 'type': 'value_error.extra'},
+        {'loc': ('spam',), 'msg': 'extra fields not permitted', 'type': 'value_error.extra'},
+    ]
+
+
+def test_disallow_mutation():
+    class Model(BaseModel):
+        a: float
+
+    model = Model(a=0.2)
+    with pytest.raises(ValueError, match='"Model" object has no field "b"'):
+        model.b = 2
+
+
+def test_extra_allowed():
+    class Model(BaseModel):
+        a: float
+
+        class Config:
+            extra = Extra.allow
+
+    model = Model(a=0.2, b=0.1)
+    assert model.b == 0.1
+
+    assert not hasattr(model, 'c')
+    model.c = 1
+    assert hasattr(model, 'c')
+    assert model.c == 1
+
+
+def test_extra_ignored():
+    class Model(BaseModel):
+        a: float
+
+        class Config:
+            extra = Extra.ignore
+
+    model = Model(a=0.2, b=0.1)
+    assert not hasattr(model, 'b')
+
+    with pytest.raises(ValueError, match='"Model" object has no field "c"'):
+        model.c = 1
 
 
 def test_set_attr():
@@ -288,7 +332,7 @@ def test_not_immutability():
 
         class Config:
             allow_mutation = True
-            allow_extra = False
+            extra = Extra.forbid
 
     m = TestModel()
     assert m.a == 10
@@ -305,7 +349,7 @@ def test_immutability():
 
         class Config:
             allow_mutation = False
-            allow_extra = False
+            extra = Extra.forbid
 
     m = TestModel()
     assert m.a == 10